+++ src/rail/core/__init__.py
@@ -18,6 +18,6 @@ except ImportError: # pragma: no cover
 
 from .stage import RailPipeline, RailStage
 #from .utilPhotometry import PhotormetryManipulator, HyperbolicSmoothing, HyperbolicMagnitudes
-from .utilStages import ColumnMapper, RowSelector, TableConverter
+from .util_stages import ColumnMapper, RowSelector, TableConverter
 from .introspection import RailEnv
 from .point_estimation import PointEstimationMixin
